+++ src/Service/Provider.php
@@ -9,7 +9,7 @@
  * @copyright 2016 (c) Tomaz Lovrec
  * @license   MIT <https://opensource.org/licenses/MIT>
  * @link      https://github.com/slaxweb/
- * @version   0.2
+ * @version   0.3
  */
 namespace SlaxWeb\Logger\Service;
 
@@ -17,6 +17,7 @@ use Pimple\Container;
 use SlaxWeb\Logger\Helper;
 use Monolog\Logger as MLogger;
 use SlaxWeb\Config\Container as Config;
+use Monolog\Registry as LoggerContainer;
 
 class Provider implements \Pimple\ServiceProviderInterface
 {
@@ -31,50 +32,67 @@ class Provider implements \Pimple\ServiceProviderInterface
      */
     public function register(Container $container)
     {
-        $container["logger.service"] = function (Container $cont) {
-            /*
-             * Check the config service has been defined and provides correct
-             * object
-             */
-            if (
-                isset($cont["config.service"]) === false
-                || get_class($cont["config.service"]) !== "SlaxWeb\\Config\\Container"
-            ) {
-                throw new \SlaxWeb\Logger\Exception\ConfigurationException(
-                    "Config component provider must be registered before you "
-                    . "can use the Logger component."
-                );
-            }
-
-            // check all config items exist
-            if ($this->_checkConfig($cont["config.service"]) === false) {
-                throw new \SlaxWeb\Logger\Exception\ConfigurationException(
-                    "Logger configuration is not complete. Make sure all "
-                    . "required settings are properly set."
-                );
-            }
-
-            // load propper handler and instantiate the Monolog\Logger
-            $handler = null;
-            switch ($cont["config.service"]["logger.loggerType"]) {
-                case Helper::L_TYPE_FILE:
-                    $handler = $cont["logger.{$cont["config.service"]["logger.loggerType"]}.service"];
-                    break;
-                default:
-                    throw new \SlaxWeb\Logger\Exception\UnknownHandlerException(
-                        "The handler you are tring to use is not known or not "
-                        . "supported."
+        $container["logger.service"] = $container->protect(
+            function (string $logger = "") use ($container)
+            {
+                /*
+                 * Check the config service has been defined and provides correct
+                 * object
+                 */
+                if (isset($container["config.service"]) === false
+                    || get_class($container["config.service"]) !== "SlaxWeb\\Config\\Container") {
+                    throw new \SlaxWeb\Logger\Exception\ConfigurationException(
+                        "Config component provider must be registered before you can use the Logger component."
                     );
+                }
+
+                $config = $container["config.service"];
+
+                // check all config items exist
+                if ($this->_checkConfig($config, $loggerName) === false) {
+                    throw new \SlaxWeb\Logger\Exception\ConfigurationException(
+                        "Logger configuration is not complete. Make sure all required settings are properly set."
+                    );
+                }
+
+                // if already in the container, return it
+                if (LoggerContainer::hasLogger($logger)) {
+                    return LoggerContainer::getInstance($logger);
+                }
+
+                // load propper handler and instantiate the Monolog\Logger
+                $handler = null;
+                $loggerType = $config["logger.loggerType"][$logger];
+                switch ($loggerType) {
+                    case Helper::L_TYPE_FILE:
+                        $container["temp.logger.property"] = [
+                            "name"  =>  $logger,
+                            "type"  =>  $type
+                        ];
+                        $handler = $container["logger.{$loggerType}.service"];
+                        break;
+                    default:
+                        throw new \SlaxWeb\Logger\Exception\UnknownHandlerException(
+                            "The handler you are tring to use is not known or not supported."
+                        );
+                }
+                $logger = new MLogger($logger);
+                $logger->pushHandler($handler);
+                return $logger;
             }
-            $logger = new MLogger($cont["config.service"]["logger.name"]);
-            $logger->pushHandler($handler);
-            return $logger;
-        };
+        );
 
         $container["logger.StreamHandler.service"] = $container->factory(
             function (Container $cont) {
+                $type = $cont["temp.logger.property"]["type"];
+                $logger = $cont["temp.logger.property"]["name"];
+                unset(
+                    $cont["temp.logger.property"]["type"],
+                    $cont["temp.logger.property"]["name"]
+                );
+
                 return new \Monolog\Handler\StreamHandler(
-                    ...$cont["config.service"]["logger.handlerArgs.{$cont["config.service"]["logger.loggerType"]}"]
+                    ...$cont["config.service"]["logger.handlerArgs.{$type}"][$logger]
                 );
             }
         );
@@ -83,16 +101,19 @@ class Provider implements \Pimple\ServiceProviderInterface
     /**
      * Check Logger configuration
      *
-     * Check that all keys that are required to instantiate the Logger component
-     * exist.
+     * Check that all keys that are required to instantiate the Logger component exist.
      *
      * @param \SlaxWeb\Config\Container $config Configuration module
+     * @param string $loggerName Name of the logger that is going to be initialized
      * @return bool
      */
-    protected function _checkConfig(Config $config): bool
+    protected function _checkConfig(Config $config, string $loggerName): bool
     {
-        return isset($config["logger.name"])
-            && isset($config["logger.loggerType"])
-            && isset($config["logger.handlerArgs.{$config["logger.loggerType"]}"]);
+        if ($loggerName === "" && ($loggerName = $config["logger.name"] ?? "") === "") {
+            return false;
+        }
+
+        return ($type = $config["logger.loggerType"][$loggerName] ?? false)
+            && $config["logger.handlerArgs.{$type}"][$loggerName] ?? false;
     }
 }