+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Changes between versions.
 
 ## Current changes
 
+* Config component no longer a requirement, but a suggestion, required if
+Factory or Service Providers are used
+* Service Provider and Factory now support only multiple loggers
+
 ## v0.2
 
 ### v0.2.0